# client/api/openapi.yaml
openapi: 3.0.1
info:
  description: REST API for the Wasp node
  title: Wasp API
  version: 0.4.0-alpha.2-402-g2adcf666b
...
servers:
- url: http://localhost:9090/
- url: https://localhost:9090/
...
paths:
...
  /chains/{chainID}:
    get:
      operationId: getChainInfo
      parameters:
      - description: ChainID (Bech32)
        in: path
        name: chainID
        required: true
        schema:
          format: string
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChainInfoResponse'
          description: Information about a specific chain
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: "Unauthorized (Wrong permissions, missing token)"
      security:
      - Authorization: []
      summary: Get information about a specific chain
      tags:
      - chains
...
  /chains/{chainID}/committee:
    get:
      operationId: getCommitteeInfo
      parameters:
      - description: ChainID (Bech32)
        in: path
        name: chainID
        required: true
        schema:
          format: string
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitteeInfoResponse'
          description: A list of all nodes tied to the chain
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: "Unauthorized (Wrong permissions, missing token)"
      security:
      - Authorization: []
      summary: Get information about the deployed committee
      tags:
      - chains
...
  /node/config:
    get:
      operationId: getConfiguration
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  example: "true"
                  format: string
                  type: string
                type: object
          description: Dumped configuration
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: "Unauthorized (Wrong permissions, missing token)"
      security:
      - Authorization: []
      summary: Return the Wasp configuration
      tags:
      - node
...
  /node/info:
    get:
      operationId: getInfo
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoResponse'
          description: Returns information about this node.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: "Unauthorized (Wrong permissions, missing token)"
      security:
      - Authorization: []
      summary: Returns private information about this node.
      tags:
      - node
...
  /node/peers/trusted:
    delete:
      operationId: distrustPeer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PeeringTrustRequest'
        description: Info of the peer to distrust
        required: true
      responses:
        "200":
          content: {}
          description: Peer was successfully distrusted
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: "Unauthorized (Wrong permissions, missing token)"
      security:
      - Authorization: []
      summary: Distrust a peering node
      tags:
      - node
      x-codegen-request-body-name: ""
    get:
      operationId: getTrustedPeers
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PeeringNodeIdentityResponse'
                type: array
          description: A list of trusted peers
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: "Unauthorized (Wrong permissions, missing token)"
      security:
      - Authorization: []
      summary: Get trusted peers
      tags:
      - node
    post:
      operationId: trustPeer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PeeringTrustRequest'
        description: Info of the peer to trust
        required: true
      responses:
        "200":
          content: {}
          description: Peer was successfully trusted
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: "Unauthorized (Wrong permissions, missing token)"
      security:
      - Authorization: []
      summary: Trust a peering node
      tags:
      - node
      x-codegen-request-body-name: ""
...
components:
  schemas:
...
    BaseToken:
      example:
        unit: TEST
        decimals: 6
        name: TestCoin
        tickerSymbol: TEST
        subunit: testies
        useMetricPrefix: true
      properties:
        decimals:
          description: The token decimals
          example: 6
          format: int32
          type: integer
          xml:
            name: Decimals
        name:
          description: The base token name
          example: TestCoin
          format: string
          type: string
          xml:
            name: Name
        subunit:
          description: The token subunit
          example: testies
          format: string
          type: string
          xml:
            name: Subunit
        tickerSymbol:
          description: The ticker symbol
          example: TEST
          format: string
          type: string
          xml:
            name: TickerSymbol
        unit:
          description: The token unit
          example: TEST
          format: string
          type: string
          xml:
            name: Unit
        useMetricPrefix:
          description: Whether or not the token uses a metric prefix
          format: boolean
          type: boolean
          xml:
            name: UseMetricPrefix
      type: object
      xml:
        name: BaseToken
...
    ChainInfoResponse:
      example:
        chainOwnerId: tst1qzjsxstc0k850jevpqu08tj0suql9u7hvh3vq3eaaem3gkx7r646zqpdn6e
        maxEventsPerReq: 50
        chainID: tst1prcw42l5u4g24tqg628d7qzh7n6m4k4ktvgayh44dyt68y930qzy2lr054v
        evmChainId: 1074
        description: description
        gasFeePolicy:
          gasFeeTokenId: gasFeeTokenId
          gasPerToken: 100
          validatorFeeShare: 0
        isActive: true
        maxBlobSize: 2000000
        maxEventSize: 2000
      properties:
        chainID:
          description: ChainID (Bech32-encoded).
          example: tst1prcw42l5u4g24tqg628d7qzh7n6m4k4ktvgayh44dyt68y930qzy2lr054v
          format: string
          type: string
          xml:
            name: ChainID
        chainOwnerId:
          description: The chain owner address (Bech32-encoded).
          example: tst1qzjsxstc0k850jevpqu08tj0suql9u7hvh3vq3eaaem3gkx7r646zqpdn6e
          format: string
          type: string
          xml:
            name: ChainOwnerID
        description:
          description: The description of the chain.
          format: string
          type: string
          xml:
            name: Description
        evmChainId:
          description: The EVM chain ID
          example: 1074
          format: int32
          type: integer
          xml:
            name: EVMChainID
        gasFeePolicy:
          $ref: '#/components/schemas/gasFeePolicy'
        isActive:
          description: Whether or not the chain is active.
          format: boolean
          type: boolean
          xml:
            name: IsActive
        maxBlobSize:
          description: The maximum contract blob size.
          example: 2000000
          format: int32
          type: integer
          xml:
            name: MaxBlobSize
        maxEventSize:
          description: The maximum event size.
          example: 2000
          format: int32
          type: integer
          xml:
            name: MaxEventSize
        maxEventsPerReq:
          description: The maximum amount of events per request.
          example: 50
          format: int32
          type: integer
          xml:
            name: MaxEventsPerReq
      type: object
      xml:
        name: ChainInfoResponse
...
    CommitteeInfoResponse:
      example:
        candidateNodes:
        - node:
            isAlive: true
            netId: netId
            publicKey: publicKey
            numUsers: 0
            isTrusted: true
          accessAPI: accessAPI
        - node:
            isAlive: true
            netId: netId
            publicKey: publicKey
            numUsers: 0
            isTrusted: true
          accessAPI: accessAPI
        committeeNodes:
        - node:
            isAlive: true
            netId: localhost:4000
            publicKey: 0x0000
            numUsers: 1
            isTrusted: true
          accessAPI: accessAPI
        - node:
            isAlive: true
            netId: localhost:4000
            publicKey: 0x0000
            numUsers: 1
            isTrusted: true
          accessAPI: accessAPI
        chainId: tst1pqm5ckama06xhkl080mmvz6l3xy8c8lulrwy7mx4ll0fc69krxfgka70j0e
        accessNodes:
        - node:
            isAlive: true
            netId: netId
            publicKey: publicKey
            numUsers: 0
            isTrusted: true
          accessAPI: accessAPI
        - node:
            isAlive: true
            netId: netId
            publicKey: publicKey
            numUsers: 0
            isTrusted: true
          accessAPI: accessAPI
        stateAddress: 0xff97a3eb5c56f6a3bc4fb729dedff9bffe37583d81d6c72ac12f2438cc94fb43
        active: true
      properties:
        accessNodes:
          description: A list of all access nodes and their peering info.
          items:
            $ref: '#/components/schemas/CommitteeNode'
          type: array
          xml:
            name: AccessNodes
            wrapped: true
        active:
          description: Whether or not the chain is active.
          example: true
          format: boolean
          type: boolean
          xml:
            name: Active
        candidateNodes:
          description: A list of all candidate nodes and their peering info.
          items:
            $ref: '#/components/schemas/CommitteeNode'
          type: array
          xml:
            name: CandidateNodes
            wrapped: true
        chainId:
          description: ChainID (Bech32-encoded).
          example: tst1pqm5ckama06xhkl080mmvz6l3xy8c8lulrwy7mx4ll0fc69krxfgka70j0e
          format: string
          type: string
          xml:
            name: ChainID
        committeeNodes:
          description: A list of all committee nodes and their peering info.
          items:
            $ref: '#/components/schemas/CommitteeNode_'
          type: array
          xml:
            name: CommitteeNodes
            wrapped: true
        stateAddress:
          example: 0xff97a3eb5c56f6a3bc4fb729dedff9bffe37583d81d6c72ac12f2438cc94fb43
          format: string
          type: string
          xml:
            name: StateAddress
      type: object
      xml:
        name: CommitteeInfoResponse
    CommitteeNode:
      example:
        node:
          isAlive: true
          netId: netId
          publicKey: publicKey
          numUsers: 0
          isTrusted: true
        accessAPI: accessAPI
      properties:
        accessAPI:
          format: string
          type: string
          xml:
            name: AccessAPI
        node:
          $ref: '#/components/schemas/PeeringNodeStatusResponse'
      type: object
      xml:
        name: CommitteeNode
    CommitteeNode_:
      example:
        node:
          isAlive: true
          netId: localhost:4000
          publicKey: 0x0000
          numUsers: 1
          isTrusted: true
        accessAPI: accessAPI
      properties:
        accessAPI:
          format: string
          type: string
          xml:
            name: AccessAPI
        node:
          $ref: '#/components/schemas/PeeringNodeStatusResponse_'
      type: object
      xml:
        name: CommitteeNode
...
    InfoResponse:
      example:
        netID: 0.0.0.0:4000
        l1Params:
          protocol:
            rentStructure:
              vByteFactorData: 1
              vByteCost: 500
              vByteFactorKey: 10
            minPowScore: 0
            tokenSupply: "2779530283277761"
            networkName: private_tangle1
            belowMaxDepth: 15
            version: 2
            bech32Hrp: tst
          maxPayloadSize: 32498
          baseToken:
            unit: TEST
            decimals: 6
            name: TestCoin
            tickerSymbol: TEST
            subunit: testies
            useMetricPrefix: true
        publicKey: 0x8fb9555a17393f0e01aadd90f8b6bc8a64b1c5c37c99915942e40b2af99e72d6
        version: 0.4.0-alpha.2-390-g5470f1925
      properties:
        l1Params:
          $ref: '#/components/schemas/L1Params'
        netID:
          description: The net id of the node
          example: 0.0.0.0:4000
          format: string
          type: string
          xml:
            name: NetID
        publicKey:
          description: The public key of the node (Hex)
          example: 0x8fb9555a17393f0e01aadd90f8b6bc8a64b1c5c37c99915942e40b2af99e72d6
          format: string
          type: string
          xml:
            name: PublicKey
        version:
          description: The version of the node
          example: 0.4.0-alpha.2-390-g5470f1925
          format: string
          type: string
          xml:
            name: Version
      type: object
      xml:
        name: InfoResponse
...
    L1Params:
      example:
        protocol:
          rentStructure:
            vByteFactorData: 1
            vByteCost: 500
            vByteFactorKey: 10
          minPowScore: 0
          tokenSupply: "2779530283277761"
          networkName: private_tangle1
          belowMaxDepth: 15
          version: 2
          bech32Hrp: tst
        maxPayloadSize: 32498
        baseToken:
          unit: TEST
          decimals: 6
          name: TestCoin
          tickerSymbol: TEST
          subunit: testies
          useMetricPrefix: true
      properties:
        baseToken:
          $ref: '#/components/schemas/BaseToken'
        maxPayloadSize:
          description: The max payload size
          example: 32498
          format: int32
          type: integer
          xml:
            name: MaxPayloadSize
        protocol:
          $ref: '#/components/schemas/ProtocolParameters'
      type: object
      xml:
        name: L1Params
...
    PeeringNodeIdentityResponse:
      example:
        netId: localhost:4000
        publicKey: 0x0000
        isTrusted: true
      properties:
        isTrusted:
          example: true
          format: boolean
          type: boolean
          xml:
            name: IsTrusted
        netId:
          description: The NetID of the peer
          example: localhost:4000
          format: string
          type: string
          xml:
            name: NetID
        publicKey:
          description: The peers public key encoded in Hex
          example: 0x0000
          format: string
          type: string
          xml:
            name: PublicKey
      type: object
      xml:
        name: PeeringNodeIdentityResponse
    PeeringNodeStatusResponse:
      example:
        isAlive: true
        netId: netId
        publicKey: publicKey
        numUsers: 0
        isTrusted: true
      properties:
        isAlive:
          description: Whether or not the peer is activated
          format: boolean
          type: boolean
          xml:
            name: IsAlive
        isTrusted:
          format: boolean
          type: boolean
          xml:
            name: IsTrusted
        netId:
          description: The NetID of the peer
          format: string
          type: string
          xml:
            name: NetID
        numUsers:
          description: The amount of users attached to the peer
          format: int32
          type: integer
          xml:
            name: NumUsers
        publicKey:
          description: The peers public key encoded in Hex
          format: string
          type: string
          xml:
            name: PublicKey
      type: object
      xml:
        name: PeeringNodeStatusResponse
    PeeringNodeStatusResponse_:
      example:
        isAlive: true
        netId: localhost:4000
        publicKey: 0x0000
        numUsers: 1
        isTrusted: true
      properties:
        isAlive:
          description: Whether or not the peer is activated
          example: true
          format: boolean
          type: boolean
          xml:
            name: IsAlive
        isTrusted:
          example: true
          format: boolean
          type: boolean
          xml:
            name: IsTrusted
        netId:
          description: The NetID of the peer
          example: localhost:4000
          format: string
          type: string
          xml:
            name: NetID
        numUsers:
          description: The amount of users attached to the peer
          example: 1
          format: int32
          type: integer
          xml:
            name: NumUsers
        publicKey:
          description: The peers public key encoded in Hex
          example: 0x0000
          format: string
          type: string
          xml:
            name: PublicKey
      type: object
      xml:
        name: PeeringNodeStatusResponse
    PeeringTrustRequest:
      example:
        netId: localhost:4000
        publicKey: 0x0000
      properties:
        netId:
          description: The NetID of the peer
          example: localhost:4000
          format: string
          type: string
          xml:
            name: NetID
        publicKey:
          description: The peers public key encoded in Hex
          example: 0x0000
          format: string
          type: string
          xml:
            name: PublicKey
      type: object
      xml:
        name: PeeringTrustRequest
    ProtocolParameters:
      example:
        rentStructure:
          vByteFactorData: 1
          vByteCost: 500
          vByteFactorKey: 10
        minPowScore: 0
        tokenSupply: "2779530283277761"
        networkName: private_tangle1
        belowMaxDepth: 15
        version: 2
        bech32Hrp: tst
      properties:
        bech32Hrp:
          description: The human readable network prefix
          example: tst
          format: string
          type: string
          xml:
            name: Bech32HRP
        belowMaxDepth:
          description: The networks max depth
          example: 15
          format: int32
          type: integer
          xml:
            name: BelowMaxDepth
        minPowScore:
          description: The minimal PoW score
          format: int32
          type: integer
          xml:
            name: MinPoWScore
        networkName:
          description: The network name
          example: private_tangle1
          format: string
          type: string
          xml:
            name: NetworkName
        rentStructure:
          $ref: '#/components/schemas/RentStructure'
        tokenSupply:
          description: The token supply
          example: "2779530283277761"
          format: string
          type: string
          xml:
            name: TokenSupply
        version:
          description: The protocol version
          example: 2
          format: int32
          type: integer
          xml:
            name: Version
      type: object
      xml:
        name: ProtocolParameters
...
    RentStructure:
      example:
        vByteFactorData: 1
        vByteCost: 500
        vByteFactorKey: 10
      properties:
        vByteCost:
          description: The virtual byte cost
          example: 500
          format: int32
          type: integer
          xml:
            name: VByteCost
        vByteFactorData:
          description: The virtual byte factor for data fields
          example: 1
          format: int32
          type: integer
          xml:
            name: VBFactorData
        vByteFactorKey:
          description: The virtual byte factor for key/lookup generating fields
          example: 10
          format: int32
          type: integer
          xml:
            name: VBFactorKey
      type: object
      xml:
        name: RentStructure
...
    ValidationError:
      properties:
        Error:
          format: string
          type: string
          xml:
            name: Error
        MissingPermission:
          format: string
          type: string
          xml:
            name: MissingPermission
      type: object
      xml:
        name: ValidationError
    gasFeePolicy:
      example:
        gasFeeTokenId: gasFeeTokenId
        gasPerToken: 100
        validatorFeeShare: 0
      properties:
        gasFeeTokenId:
          description: The gas fee token id. Empty if base token.
          format: string
          type: string
          xml:
            name: GasFeeTokenID
        gasPerToken:
          description: The amount of gas per token.
          example: 100
          format: int64
          type: integer
          xml:
            name: GasPerToken
        validatorFeeShare:
          description: The validator fee share.
          format: int32
          type: integer
          xml:
            name: ValidatorFeeShare
      type: object
      xml:
        name: gasFeePolicy
...
  securitySchemes:
    Authorization:
      description: JWT Token
      in: header
      name: Authorization
      type: apiKey
...